Default Display

+++ yln/Main/Views/ContentView.swift
@@ -13,20 +13,26 @@ struct ContentView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-            List {
-                ForEach(listViewModel.items) { item in
-//                    Text("Hi")
-                    ListRowView(item: item)
-                        .onTapGesture {
-                            withAnimation(.linear){
-                                listViewModel.updateItem(item: item)
+        ZStack{
+            if listViewModel.items.isEmpty {
+                Text("No items")
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+    //                    Text("Hi")
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
                             }
-                        }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
                 }
-                .onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
+                .listStyle(PlainListStyle())
             }
-            .listStyle(PlainListStyle())
+        }
             .navigationTitle("Todo List📝")
             .navigationBarItems(
                 leading: EditButton(),
